Fix constructor  of transaction creator
